# contract/openapi.yml
openapi: 3.0.0
info:
  title: IT Corpo API
  version: 2.0.0
  description: >
    API for managing IT corporation resources.

    For full documentation, please visit [the readme](https://github.com/ducin-public/itcorpo-api).
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /benefits/services:
    get:
      summary: List all available benefit services
      operationId: getBenefitServices
      tags:
        - Benefits
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BenefitService'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /benefits:
    get:
      summary: List all benefits subscriptions
      operationId: getBenefitSubscriptions
      tags:
        - Benefits
        - Pagination
        - Search
      parameters:
        - $ref: '#/components/parameters/BenefitServiceName'
        - $ref: '#/components/parameters/BenefitCategories'
        - $ref: '#/components/parameters/BenefitEmployeeId'
        - $ref: '#/components/parameters/BenefitFeeFrom'
        - $ref: '#/components/parameters/BenefitFeeTo'
        - $ref: '#/components/parameters/BenefitStatus'
        - $ref: '#/components/parameters/PageNumber'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BenefitSubscription'
        '400':
          description: >
            Invalid benefit subscriptions search criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    post:
      summary: Create a new benefit
      operationId: createBenefit
      tags:
        - Benefits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BenefitSubscriptionInput'
      responses:
        '201':
          description: Benefit created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenefitSubscription'
        '400':
          description: >
            Invalid benefit subscription input request body
            @see {@link BenefitSubscriptionInput}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Benefit already exists for this employee and service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /benefits/count:
    get:
      summary: Get total number of benefits
      operationId: getBenefitsCount
      tags:
        - Benefits
        - Search
      parameters:
        - $ref: '#/components/parameters/BenefitServiceName'
        - $ref: '#/components/parameters/BenefitCategories'
        - $ref: '#/components/parameters/BenefitEmployeeId'
        - $ref: '#/components/parameters/BenefitFeeFrom'
        - $ref: '#/components/parameters/BenefitFeeTo'
        - $ref: '#/components/parameters/BenefitStatus'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: integer
                example: 42
        '400':
          description: >
            Invalid benefit subscriptions search criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /benefits/{benefitId}:
    get:
      summary: Get benefit by ID
      operationId: getBenefitSubscriptionById
      tags:
        - Benefits
      parameters:
        - name: benefitId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenefitSubscription'
        '404':
          description: Benefit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    
    put:
      summary: Update benefit
      operationId: updateBenefit
      tags:
        - Benefits
      parameters:
        - name: benefitId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BenefitSubscriptionInput'
      responses:
        '200':
          description: Benefit updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenefitSubscription'
        '400':
          description: >
            Invalid benefit subscription input request body
            @see {@link BenefitSubscriptionInput}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Benefit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    patch:
      summary: Cancel or renew benefit subscription
      operationId: updateBenefitSubscriptionStatus
      tags:
        - Benefits
      parameters:
        - name: benefitId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - operation
              properties:
                operation:
                  type: string
                  enum: ["CANCEL", "RENEW"]
      responses:
        '200':
          description: Benefit subscription status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenefitSubscription'
        '400':
          description: >
            Invalid benefit subscription status update
            @see {@link BenefitSubscriptionInput}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Benefit subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /benefits/{benefitId}/charges:
    get:
      summary: List all benefit charges for a specific subscription
      operationId: getBenefitSubscriptionCharges
      tags:
        - Benefits
        - Pagination
      parameters:
        - name: benefitId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/BenefitProviderServiceCode'
        - $ref: '#/components/parameters/BenefitChargeStatus'
        - $ref: '#/components/parameters/BenefitBillingPeriodFrom'
        - $ref: '#/components/parameters/BenefitBillingPeriodTo'
        - $ref: '#/components/parameters/PageNumber'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BenefitCharge'
        '400':
          description: 
            Invalid benefit charges search criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Benefit subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /benefits/charges:
    get:
      summary: List all benefit charges
      operationId: getBenefitCharges
      tags:
        - Benefits
        - Pagination
        - Search
      parameters:
        - $ref: '#/components/parameters/BenefitSubscriptionId'
        - $ref: '#/components/parameters/BenefitEmployeeId'
        - $ref: '#/components/parameters/BenefitProviderServiceCode'
        - $ref: '#/components/parameters/BenefitChargeStatus'
        - $ref: '#/components/parameters/BenefitBillingPeriodFrom'
        - $ref: '#/components/parameters/BenefitBillingPeriodTo'
        - $ref: '#/components/parameters/PageNumber'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BenefitCharge'
        '400':
          description: >
            Invalid benefit charges search criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
...
  responses:
...
    InternalServerError:
      description: An internal server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: The service is temporarily unavailable. Maybe there is maintenance?
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Money:
      type: number
      description: Monetary value in EUR
...
    Email:
      type: string
      format: email
      description: Email address string
    Phone:
      type: string
      description: Phone number string
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: An application-level error code (**not** HTTP status code)
        message:
          type: string
          description: Text description of the error that has occurred
      required:
        - message
...
    BenefitCategory:
      type: string
      enum:
        - "HEALTHCARE"
        - "SPORT_WELLNESS"
        - "LUNCH_FOOD"
        - "CULTURE_RECREATION"
      example: "HEALTHCARE"
    BenefitService:
      type: object
      required:
        - code
        - name
        - category
        - provider
        - description
        - availableCountries
      properties:
        code:
          type: string
        name:
          type: string
          description: Display name of the service
          example: "Medicover Healthcare Premium"
        category:
          $ref: '#/components/schemas/BenefitCategory'
        provider:
          type: object
          required:
            - name
            - website
            - contactEmail
            - supportPhone
          properties:
            name: 
              type: string
              example: "Medicover"
            website:
              type: string
              format: uri
              example: "https://medicover.com"
            contactEmail:
              $ref: '#/components/schemas/Email'
            supportPhone:
              $ref: '#/components/schemas/Phone'
            description:
              type: string
        description:
          type: string
          example: "Premium healthcare package including dental and specialist care"
        availableCountries:
          type: array
          items:
            type: string
          example: ["PL", "DE", "NL"]
        details:
          type: string
          description: Combined features and limitations in a formatted text
          example: "Features:\n- Dental care\n- 24/7 hotline\n- Mobile app\n\nLimitations:\n- 14 days waiting period\n- Excludes pre-existing conditions"
        cancellationPolicy:
          type: string
          example: "1 month notice required"
    BenefitSubscription:
      type: object
      required:
        - id
        - beneficiary
        - category
        - country
        - city
        - service
        - monthlyFee
        - subscribedAtDate
      properties:
        id:
          type: string
        service:
          type: object
          required:
            - name
            - provider
          properties:
            name:
              type: string
            provider:
              type: string
        beneficiary:
          type: object
          required:
            - name
            - email
          properties:
            name:
              type: string
            email:
              type: string
              format: email
        category:
          $ref: '#/components/schemas/BenefitCategory'
        country:
          type: string
        city:
          type: string
        monthlyFee:
          $ref: '#/components/schemas/Money'
        subscribedAtDate:
          type: string
          format: date
        cancelledAtDate:
          type: string
          format: date
      example:
        id: "60965cfccf2844a6"
        beneficiary:
          name: "Eva Koster"
          email: "evko@softix.nl"
        city: "Utrecht"
        country: "Netherlands"
        service:
          name: "MultiSport Active Plus"
          provider: "Benefit Systems"
        monthlyFee: 250
        subscribedAtDate: "2016-01-01"
        cancelledAtDate: "2016-05-31"
    BenefitSubscriptionInput:
      type: object
      required:
        - beneficiary
        - country
        - city
        - service
        - monthlyFee
        - subscribedAtDate
      properties:
        beneficiary:
          type: object
          required:
            - name
            - email
          properties:
            name:
              type: string
            email:
              type: string
              format: email
        country:
          type: string
        city:
          type: string
        service:
          type: string
        monthlyFee:
          $ref: '#/components/schemas/Money'
        subscribedAtDate:
          type: string
          format: date
        cancelledAtDate:
          type: string
          format: date
    BenefitSubscriptionSearchStatus:
      type: string
      enum: ["ALL", "ACTIVE", "CANCELLED"]
      description: Status filter for benefit subscriptions
      example: "ACTIVE"
    BenefitChargeStatus:
      type: string
      description: Payment status of a benefit charge
      enum:
        - "PENDING"
        - "PAID"
        - "OVERDUE"
        - "CANCELLED"
        - "REFUNDED"
    BenefitCharge:
      type: object
      required:
        - id
        - employeeId
        - subscriptionId
        - providerServiceCode
        - billingPeriodStart
        - billingPeriodEnd
        - amount
        - status
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: integer
        subscriptionId:
          type: string
        providerServiceCode:
          type: string
        billingPeriodStart:
          type: string
          format: date
        billingPeriodEnd:
          type: string
          format: date
        amount:
          $ref: '#/components/schemas/Money'
        status:
          $ref: '#/components/schemas/BenefitChargeStatus'
...
  parameters:
    PageNumber:
      name: _page
      in: query
      required: false
      description: Paginate data collections
      schema:
        type: number
        description: Page number to retrieve
        example: 1

    # Benefit parameters
    BenefitServiceName:
      name: serviceName
      in: query
      required: false
      description: Filter benefits by service name
      schema:
        type: string
        description: Filter benefits by service name
        example: "MultiSport"
    BenefitCategories:
      name: categories
      in: query
      required: false
      description: Filter benefits by categories
      schema:
        type: string
        description: >
          Comma-separated list of category codes to filter by.
          The search result will return benefits that match any of the provided categories (`ANY`).
        example: "HEALTHCARE,SPORT_WELLNESS"
    BenefitEmployeeId:
      name: employeeId
      in: query
      required: false
      description: Filter benefits by employee ID
      schema:
        type: string
        description: The employee whom this benefit is subscribed to
        example: "91720"
    BenefitFeeFrom:
      name: feeFrom
      in: query
      required: false
      description: Minimum monthly fee amount
      schema:
        type: string
        description: Minimum monthly fee amount
        example: "100"
    BenefitFeeTo:
      name: feeTo
      in: query
      required: false
      description: Maximum monthly fee amount
      schema:
        type: string
        description: Maximum monthly fee amount
        example: "500.50"
    BenefitStatus:
      name: status
      in: query
      required: false
      description: Filter benefits by status
      schema:
        $ref: '#/components/schemas/BenefitSubscriptionSearchStatus'
        example: "ACTIVE"

    # Benefit Charge parameters
    BenefitSubscriptionId:
      name: subscriptionId
      in: query
      required: false
      description: Filter charges by subscription ID
      schema:
        type: string
        description: Filter charges by subscription ID
        example: 'zc9b3b4c-1b1d-4b3e-8b3b-4c1b1d4b3e8b'
    BenefitProviderServiceCode:
      name: providerServiceCode
      in: query
      required: false
      description: Filter charges by provider service code
      schema:
        type: string
        description: Filter charges by provider service code
        example: "MEDICOVER_PREMIUM"
    BenefitChargeStatus:
      name: status
      in: query
      required: false
      description: Filter charges by status
      schema:
        $ref: '#/components/schemas/BenefitChargeStatus'
        description: Filter charges by status
        example: "PAID"
    BenefitBillingPeriodFrom:
      name: billingPeriodFrom
      in: query
      required: false
      description: Filter charges with billing period starting from this date
      schema:
        type: string
        format: date
        description: Filter charges with billing period starting from this date
        example: "2023-01-01"
    BenefitBillingPeriodTo:
      name: billingPeriodTo
      in: query
      required: false
      description: Filter charges with billing period ending before this date
      schema:
        type: string
        format: date
        description: Filter charges with billing period ending before this date
        example: "2023-12-31"

    # Employee parameters
...